review: update sha helper imports after module split
